Plugin Refactor
- Moved majority of the app setup into plugins
