+++ src/plugins/main_menu_plugin.rs
@@ -1,0 +1,1 @@
+//use bevy::prelude::*;
