+++ src/plugins/mod.rs
@@ -1,0 +1,4 @@
+pub mod main_menu_plugin;
+pub mod map_plugin;
+pub mod player_plugin;
+pub mod pregame_plugin;
